Make the failing test for #205 exclusive to Scala 2.12+
