+++ core/src/test/scala-2.12+/pureconfig/CoproductConvertersSuite2_12.scala
@@ -1,0 +1,22 @@
+package pureconfig
+
+import com.typesafe.config.ConfigFactory
+
+class CoproductConvertersSuite2_12 extends BaseSuite {
+
+  behavior of "ConfigConvert"
+
+  it should "read disambiguation information on sealed families with the cases nested in the companion" in {
+    import CarMaker._
+    val conf = ConfigFactory.parseString("{ type = bmw }")
+    ConfigConvert[CarMaker].from(conf.root()) shouldEqual Right(BMW)
+  }
+}
+
+sealed trait CarMaker
+
+object CarMaker {
+  case object Mercedes extends CarMaker
+  case object BMW extends CarMaker
+  case object Tesla extends CarMaker
+}
